+++ test/js/main.js
@@ -1,3 +1,11 @@
+<template>
+div
+
+</template>
+
+
+
+
 var app = new Vue({
   el: '#app',
   data: {
