+++ test/page.html
@@ -17,10 +17,10 @@
 	<body>
 		<div id="app">
 			<div class="container-fluid">
-				<div class="row ">
+				<div class="row  ">
 					<div class="col-6 p-0 " style="height:100vh;">
 						<form id="CodeForm" class="h-100">
-							<textarea id="CodeInput" class="h-100 w-100 black border-none"   name="text" placeholder="// поехали!"></textarea>
+							<textarea id="CodeInput" class="h-100 w-100 black border-none text-white"   name="text" placeholder="// поехали!"></textarea>
 							<button type="button"class="sub" name="button">√</button>
 							<input class="sub" type="submit" value="√">
 						</form>
@@ -49,8 +49,8 @@
 		<!-- MDB core JavaScript -->
 		<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/mdbootstrap/4.15.0/js/mdb.min.js"></script>
 		<script src="https://cdn.jsdelivr.net/npm/vue/dist/vue.js"></script>
-		<script src="./main.js"></script>
-		<script src="interpretator.js"></script>
+		<script src="./js/main.js"></script>
+		<script src="js/interpretator.js"></script>
 	</body>
 
 </html>
